fix(webpack-exodus): Fixed v2.3000.x and v2.24
